+++ Bài Tập Lớn/DAL/ChiTietPhienDAL.cs
@@ -67,7 +67,7 @@ namespace Bài_Tập_Lớn.DAL
             }
         }
 
-        public decimal TinhTongTienTheoPhien(string maPhien)
+        public double TinhTongTienTheoPhien(string maPhien)
         {
             string sql = @"
                 SELECT ISNULL(SUM(so_luong * don_gia),0)
@@ -78,7 +78,7 @@ namespace Bài_Tập_Lớn.DAL
             {
                 using (IDbConnection conn = DBConnection.Instance.GetConnection())
                 {
-                    return conn.ExecuteScalar<decimal>(
+                    return conn.ExecuteScalar<double>(
                         sql,
                         new { MaPhien = maPhien }
                     );
